--- requirements.txt
@@ -1,181 +1,0 @@
-aiohttp==3.8.5
-aiosignal==1.3.1
-altair==5.1.2
-ansi2html==1.8.0
-anyio==4.0.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.2.3
-asttokens==2.4.0
-async-lru==2.0.4
-async-timeout==4.0.3
-attrs==23.1.0
-Babel==2.12.1
-backcall==0.2.0
-beautifulsoup4==4.12.2
-bleach==6.0.0
-blinker==1.6.3
-bokeh==2.4.3
-cachetools==5.3.1
-category-encoders==2.6.2
-certifi==2023.7.22
-cffi==1.15.1
-chardet==5.2.0
-charset-normalizer==3.2.0
-click==8.1.7
-cloudpickle==2.2.1
-colorama==0.4.6
-comm==0.1.4
-connected-components-3d==3.12.2
-contourpy==1.1.0
-cycler==0.11.0
-dash==2.14.1
-dash-bootstrap-components==1.5.0
-dash-core-components==2.0.0
-dash-html-components==2.0.0
-dash-table==5.0.0
-dask==2023.9.0
-dataprep==0.4.5
-debugpy==1.6.7.post1
-decorator==5.1.1
-defusedxml==0.7.1
-exceptiongroup==1.1.3
-executing==0.8.3
-fastjsonschema==2.18.0
-Flask==2.2.5
-Flask-Cors==3.0.10
-fonttools==4.40.0
-fqdn==1.5.1
-frozenlist==1.4.0
-fsspec==2023.9.0
-gitdb==4.0.10
-GitPython==3.1.37
-idna==3.4
-importlib-metadata==6.8.0
-ipykernel==6.25.2
-ipython==8.15.0
-ipython-genutils==0.2.0
-ipywidgets==7.8.0
-isoduration==20.11.0
-itsdangerous==2.1.2
-jedi==0.19.0
-Jinja2==3.0.3
-joblib==1.3.1
-json5==0.9.14
-jsonpath-ng==1.5.3
-jsonpointer==2.4
-jsonschema==4.19.0
-jsonschema-specifications==2023.7.1
-jupyter-events==0.7.0
-jupyter-lsp==2.2.0
-jupyter_client==8.3.1
-jupyter_core==5.3.1
-jupyter_server==2.7.3
-jupyter_server_terminals==0.4.4
-jupyterlab==4.0.5
-jupyterlab-pygments==0.2.2
-jupyterlab-widgets==1.1.5
-jupyterlab_server==2.24.0
-kiwisolver==1.4.4
-locket==1.0.0
-markdown-it-py==3.0.0
-MarkupSafe==2.1.3
-matplotlib==3.7.2
-matplotlib-inline==0.1.6
-mdurl==0.1.2
-Metaphone==0.6
-mistune==3.0.1
-multidict==6.0.4
-mysql-connector==2.2.9
-mysql-connector-python==8.3.0
-nbclient==0.8.0
-nbconvert==7.8.0
-nbformat==5.9.2
-nest-asyncio==1.5.7
-nltk==3.8.1
-notebook==7.0.3
-notebook_shim==0.2.3
-numpy==1.25.0
-opencv-python==4.8.1.78
-overrides==7.4.0
-packaging==23.1
-pandas==1.5.3
-pandocfilters==1.5.0
-parso==0.8.3
-partd==1.4.0
-patsy==0.5.3
-pickleshare==0.7.5
-Pillow==10.0.0
-platformdirs==3.10.0
-plotly @ file:///opt/conda/envs/env/conda-bld/plotly_1694788482196/work/packages/python/plotly/dist/plotly-5.17.0.tar.gz
-ply==3.11
-prometheus-client==0.17.1
-prompt-toolkit==3.0.39
-protobuf==4.24.4
-psutil==5.9.5
-pure-eval==0.2.2
-pyarrow==13.0.0
-pycparser==2.21
-pydantic==1.10.12
-pydeck==0.8.1b0
-pydot==1.4.2
-Pygments==2.16.1
-pyparsing==3.0.9
-python-crfsuite==0.9.8
-python-dateutil==2.8.2
-python-json-logger==2.0.7
-python-stdnum==1.19
-pytz==2023.3
-pywin32==306
-pywinpty==2.0.11
-PyYAML==6.0.1
-pyzmq==25.1.1
-rapidfuzz==2.15.1
-referencing==0.30.2
-regex==2021.11.10
-requests==2.31.0
-retrying==1.3.4
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rich==13.6.0
-rpds-py==0.10.2
-scikit-learn==1.3.0
-scipy==1.11.1
-seaborn==0.12.2
-Send2Trash==1.8.2
-six==1.16.0
-smmap==5.0.1
-sniffio==1.3.0
-soupsieve==2.5
-SQLAlchemy==1.3.24
-stack-data==0.5.1
-statsmodels==0.14.0
-streamlit==1.27.2
-tenacity @ file:///C:/b/abs_ddkoa9nju6/croot/tenacity_1682972298929/work
-terminado==0.17.1
-threadpoolctl==3.1.0
-tinycss2==1.2.1
-toml==0.10.2
-tomli==2.0.1
-toolz==0.12.0
-tornado==6.3.3
-tqdm==4.66.1
-traitlets==5.9.0
-typing_extensions==4.7.1
-tzdata==2023.3
-tzlocal==5.1
-uri-template==1.3.0
-urllib3==2.0.4
-validators==0.22.0
-varname==0.8.3
-watchdog==3.0.0
-waterfallcharts==3.8
-wcwidth==0.2.6
-webcolors==1.13
-webencodings==0.5.1
-websocket-client==1.6.2
-Werkzeug==2.3.7
-widgetsnbextension==3.6.5
-wordcloud==1.9.2
-yarl==1.9.2
-zipp==3.16.2
