+++ spb/app.py
@@ -1,0 +1,38 @@
+from flask import Flask
+from flask import render_template
+from flask import request
+from flask import redirect
+from flask import url_for
+import re
+from datetime import datetime
+import mysql.connector
+from mysql.connector import FieldType
+import connect
+
+app = Flask(__name__)
+
+dbconn = None
+connection = None
+
+def getCursor():
+    global dbconn
+    global connection
+    connection = mysql.connector.connect(user=connect.dbuser, \
+    password=connect.dbpass, host=connect.dbhost, \
+    database=connect.dbname, autocommit=True)
+    dbconn = connection.cursor()
+    return dbconn
+
+@app.route("/")
+def home():
+    return redirect("/currentjobs")
+
+@app.route("/currentjobs")
+def currentjobs():
+    connection = getCursor()
+    connection.execute("SELECT job_id,customer,job_date FROM job where completed=0;")
+    jobList = connection.fetchall()
+    return render_template("currentjoblist.html", job_list = jobList)    
+
+
+
