+++ spb/static/Thumbs.db
