--- templates/administrator.html
@@ -1,18 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Administrator{% endblock %}
-
-
-{% block header %}
-<div class="header">
-    <nav>
-        <ul>
-            <li><a href="{{ url_for('customer') }}">Customer</a></li>
-            <li><a href="{{ url_for('service') }}">Service</a></li>
-            <li><a href="{{ url_for('part') }}">Part</a></li>
-            <li><a href="{{ url_for('billinghistory') }}">Billing History</a></li>
-        </ul>
-    </nav>
-</div>
-{% endblock %}
-
-
