--- templates/currentjoblist.html
@@ -1,41 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Current Job{% endblock %}
-
-{% block header %}
-<div class="header">
-    <nav>
-        <ul>
-            <li><a href="{{ url_for('administrator') }}">Administrator</a></li>
-        </ul>
-    </nav>
-</div>
-{% endblock %}
-
-{% block content %}
-<style>
-    .table-container {
-    max-width: 1200px;
-    margin: 0 auto;
-    }
-
-
-</style>
-<div class="table-container">
-    <h2>Current Job List</h2>
-    <!-- code to display list of jobs-->
-    <table class="table">
-        <tr>
-            <th>Job ID</th>
-            <th>Customer Name</th>
-            <th>Date</th>
-        </tr>
-        {% for job in job_list %}
-        <tr>
-            {% for field in job %}
-            <td>{{field}}</td>
-            {% endfor %}
-        </tr>
-        {% endfor %}
-    </table>
-</div>
-{% endblock %}
