--- templates/service.html
@@ -1,2 +1,0 @@
-{% extends "administrator.html" %}
-{% block title %}Service{% endblock %}
